Reintroduced MyCanvas() back into Drawing.js, removed MyCanvas.js
